[pandora car alarm] Fix mileage type

+++ src/main/java/org/openhab/binding/pandoracaralarmsystem/internal/api/record/StatRecord.java
@@ -116,13 +116,13 @@ public class StatRecord {
     /**
      * The Mileage.
      */
-    public Long mileage;
+    public Float mileage;
 
     /**
      * The Mileage can.
      */
     @SerializedName("mileage_CAN")
-    public Long mileageCAN;
+    public Float mileageCAN;
 
     /**
      * The Metka.
